# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "uiscloud_weknora 지식베이스 관리 시스템 API 문서",
        "title": "uiscloud_weknora API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "uiscloud_weknora Github",
            "url": "https://github.com/rockgis/uiscloud_weknora"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/messages/{session_id}/load": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "세션의 메시지 히스토리를 로드합니다. 페이지네이션과 시간 필터를 지원합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "메시지"
                ],
                "summary": "메시지 히스토리 로드",
                "parameters": [
                    {
                        "type": "string",
                        "description": "세션 ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "반환 수량",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "이 시간 이전의 메시지 (RFC3339Nano 형식)",
                        "name": "before_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "메시지 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/messages/{session_id}/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "세션에서 지정된 메시지를 삭제합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "메시지"
                ],
                "summary": "메시지 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "세션 ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "메시지 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Tencent_uiscloud_weknora_internal_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010,
                2000,
                2001,
                2002,
                2003,
                2004,
                2100,
                2101,
                2102,
                2103
            ],
            "x-enum-varnames": [
                "ErrBadRequest",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrNotFound",
                "ErrMethodNotAllowed",
                "ErrConflict",
                "ErrTooManyRequests",
                "ErrInternalServer",
                "ErrServiceUnavailable",
                "ErrTimeout",
                "ErrValidation",
                "ErrTenantNotFound",
                "ErrTenantAlreadyExists",
                "ErrTenantInactive",
                "ErrTenantNameRequired",
                "ErrTenantInvalidStatus",
                "ErrAgentMissingThinkingModel",
                "ErrAgentMissingAllowedTools",
                "ErrAgentInvalidMaxIterations",
                "ErrAgentInvalidTemperature"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer {token} 형식의 JWT 토큰을 입력하세요",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}